add RoomMessageException
